Harden faucet scheduler and release checks
Scope faucet persistence batches away from async calls, add retry-once
governance staking-subaccount lookup, and cover the async/state invariant
with scheduler tests. Strengthen reproducible-build and release validation
by checking rustup-init, scanning production DIDs for debug surfaces, and
hardening build-script DID parsing.

Also document intentional best-effort retry, debug guard, memo-principal,
and neuron claim/refresh trade-offs, and add stable-state compatibility
coverage.

Why: reduce scheduler durability risk, align transient-call retry behavior,
prevent release/debug artifact confusion, and make accepted operational
trade-offs explicit before deployment.

+++ jupiter-disburser/src/lib.rs
@@ -121,7 +121,7 @@ fn validate_config(cfg: &crate::state::Config) {
 
 #[ic_cdk::init]
 fn init(args: InitArgs) {
-    let now_secs = (ic_cdk::api::time() / 1_000_000_000) as u64;
+    let now_secs = ic_cdk::api::time() / 1_000_000_000;
 
     let cfg = crate::state::Config {
         neuron_id: args.neuron_id,
@@ -176,7 +176,7 @@ pub(crate) fn apply_upgrade_args_to_state(st: &mut State, args: Option<UpgradeAr
 
 #[ic_cdk::post_upgrade]
 fn post_upgrade(args: Option<UpgradeArgs>) {
-    let now_secs = (ic_cdk::api::time() / 1_000_000_000) as u64;
+    let now_secs = ic_cdk::api::time() / 1_000_000_000;
     crate::state::init_stable_storage();
     let mut st = crate::state::restore_state_from_stable().expect("stable state missing during disburser post_upgrade");
     let actions = apply_upgrade_args_to_state(&mut st, args, now_secs);
