+++ jupiter-faucet/src/lib.rs
@@ -144,7 +144,7 @@ fn decode_subaccount_opt(v: Option<Vec<u8>>) -> Result<Option<[u8; 32]>, String>
 
 #[ic_cdk::init]
 fn init(args: InitArgs) {
-    let now_secs = (ic_cdk::api::time() / 1_000_000_000) as u64;
+    let now_secs = ic_cdk::api::time() / 1_000_000_000;
 
     let cfg = crate::state::Config {
         staking_account: args.staking_account,
@@ -214,7 +214,7 @@ pub(crate) fn apply_upgrade_args_to_state(st: &mut State, args: Option<UpgradeAr
 
 #[ic_cdk::post_upgrade]
 fn post_upgrade(args: Option<UpgradeArgs>) {
-    let now_secs = (ic_cdk::api::time() / 1_000_000_000) as u64;
+    let now_secs = ic_cdk::api::time() / 1_000_000_000;
     crate::state::init_stable_storage();
     let mut st = crate::state::restore_state_from_stable().expect("stable state missing during faucet post_upgrade");
     let actions = apply_upgrade_args_to_state(&mut st, args, now_secs);
